chore: Split parser into separate modules
